+++ common/enhancedportals/lib/Localization.java
@@ -1,5 +1,6 @@
 package enhancedportals.lib;
 
+import net.minecraft.util.StatCollector;
 import cpw.mods.fml.common.registry.LanguageRegistry;
 
 public class Localization
@@ -12,11 +13,6 @@ public class Localization
 
     public static String[] Locales = new String[] { "en_GB", "en_US", "it_IT" };
 
-    public static String getLocalizedString(String string)
-    {
-        return LanguageRegistry.instance().getStringLocalization(string);
-    }
-
     public static void loadLocales()
     {
         for (String str : Locales)
@@ -27,7 +23,13 @@ public class Localization
 
     public static String localizeString(String str)
     {
-        // TODO
-        return str;
+        String str2 = LanguageRegistry.instance().getStringLocalization(str);
+        
+        if (str2.length() == 0)
+        {
+            str2 = StatCollector.translateToLocal(str);
+        }
+        
+        return str2;
     }
 }